refactor: RoomService에서 CommentService를 참조하는 문제 해결
