+++ backend/src/main/java/com/digginroom/digginroom/service/CommentService.java
@@ -2,13 +2,16 @@ package com.digginroom.digginroom.service;
 
 import static com.digginroom.digginroom.exception.CommentException.NotOwnerException;
 
-import com.digginroom.digginroom.service.dto.CommentRequest;
-import com.digginroom.digginroom.service.dto.CommentResponse;
-import com.digginroom.digginroom.service.dto.CommentsResponse;
 import com.digginroom.digginroom.domain.comment.Comment;
 import com.digginroom.digginroom.domain.member.Member;
 import com.digginroom.digginroom.exception.CommentException.NotSameRoomException;
+import com.digginroom.digginroom.exception.RoomException.NotFoundException;
 import com.digginroom.digginroom.repository.CommentRepository;
+import com.digginroom.digginroom.repository.MemberRepository;
+import com.digginroom.digginroom.repository.RoomRepository;
+import com.digginroom.digginroom.service.dto.CommentRequest;
+import com.digginroom.digginroom.service.dto.CommentResponse;
+import com.digginroom.digginroom.service.dto.CommentsResponse;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -20,33 +23,52 @@ import org.springframework.transaction.annotation.Transactional;
 public class CommentService {
 
     private final CommentRepository commentRepository;
+    private final MemberRepository memberRepository;
+    private final RoomRepository roomRepository;
 
-    public CommentsResponse getRoomComments(final Long roomId, final Member member) {
+    public CommentsResponse getRoomComments(final Long roomId, final Long memberId) {
+        Member member = memberRepository.getMemberById(memberId);
         List<Comment> comments = commentRepository.findCommentsByRoomId(roomId);
         return new CommentsResponse(comments.stream()
                 .map(comment -> CommentResponse.of(comment, comment.isOwner(member)))
                 .toList());
     }
 
-    public CommentResponse comment(final Long roomId, final Member member, final CommentRequest request) {
+    public CommentResponse comment(final Long roomId, final Long memberId, final CommentRequest request) {
+        validateExistRoom(roomId);
+        Member member = memberRepository.getMemberById(memberId);
+
         Comment comment = new Comment(roomId, request.comment(), member);
         commentRepository.save(comment);
 
         return CommentResponse.of(comment, comment.isOwner(member));
     }
 
-    public void delete(final Long roomId, final Member member, final Long commentId) {
+    public void validateExistRoom(final Long roomId) {
+        if (!roomRepository.existsById(roomId)) {
+            throw new NotFoundException(roomId);
+        }
+    }
+
+    public void delete(final Long roomId, final Long memberId, final Long commentId) {
+        Member member = memberRepository.getMemberById(memberId);
         Comment comment = commentRepository.getCommentById(commentId);
 
         validateWriteable(comment, member, roomId);
         commentRepository.delete(comment);
     }
 
-    public Comment update(final Member member, final Long roomId, final Long commentId, final CommentRequest request) {
+    public CommentResponse update(
+            final Long roomId,
+            final Long memberId,
+            final Long commentId,
+            final CommentRequest request
+    ) {
+        Member member = memberRepository.getMemberById(memberId);
         Comment comment = commentRepository.getCommentById(commentId);
         validateWriteable(comment, member, roomId);
         comment.updateComment(request.comment());
-        return comment;
+        return CommentResponse.of(comment, comment.isOwner(member));
     }
 
     private void validateWriteable(final Comment comment, final Member member, final Long roomId) {
