+++ backend/src/main/java/com/digginroom/digginroom/service/RoomService.java
@@ -1,15 +1,11 @@
 package com.digginroom.digginroom.service;
 
 import com.digginroom.digginroom.domain.Genre;
-import com.digginroom.digginroom.domain.comment.Comment;
 import com.digginroom.digginroom.domain.member.Member;
 import com.digginroom.digginroom.domain.room.Room;
 import com.digginroom.digginroom.exception.RoomException.NotFoundException;
 import com.digginroom.digginroom.repository.MemberRepository;
 import com.digginroom.digginroom.repository.RoomRepository;
-import com.digginroom.digginroom.service.dto.CommentRequest;
-import com.digginroom.digginroom.service.dto.CommentResponse;
-import com.digginroom.digginroom.service.dto.CommentsResponse;
 import com.digginroom.digginroom.service.dto.RoomResponse;
 import com.digginroom.digginroom.service.dto.RoomsResponse;
 import com.digginroom.digginroom.service.dto.TrackResponse;
@@ -26,7 +22,6 @@ public class RoomService {
 
     private final RoomRepository roomRepository;
     private final MemberRepository memberRepository;
-    private final CommentService commentService;
 
     @Transactional(readOnly = true)
     public RoomResponse recommend(final Long memberId) {
@@ -98,41 +93,5 @@ public class RoomService {
         Member member = memberRepository.getMemberById(memberId);
 
         member.undislike(room);
-    }
-
-    public CommentsResponse findRoomComments(final Long roomId, final Long loginMemberId) {
-        validateExistRoom(roomId);
-        Member member = memberRepository.getMemberById(loginMemberId);
-        return commentService.getRoomComments(roomId, member);
-    }
-
-    public void validateExistRoom(final Long roomId) {
-        if (!roomRepository.existsById(roomId)) {
-            throw new NotFoundException(roomId);
-        }
-    }
-
-    public CommentResponse comment(final Long roomId, final Long memberId, final CommentRequest request) {
-        validateExistRoom(roomId);
-        Member member = memberRepository.getMemberById(memberId);
-        return commentService.comment(roomId, member, request);
-    }
-
-    public void deleteComment(final Long roomId, final Long memberId, final Long commentId) {
-        validateExistRoom(roomId);
-        Member member = memberRepository.getMemberById(memberId);
-        commentService.delete(roomId, member, commentId);
-    }
-
-    public CommentResponse updateComment(
-            final Long roomId,
-            final Long memberId,
-            final Long commentId,
-            final CommentRequest request
-    ) {
-        validateExistRoom(roomId);
-        Member member = memberRepository.getMemberById(memberId);
-        Comment updateComment = commentService.update(member, roomId, commentId, request);
-        return CommentResponse.of(updateComment, updateComment.isOwner(member));
     }
 }
